Rename detailedTvl path to tvl (#2374)
