+++ packages/frontend/src/pages/bridges/summary/index.tsx
@@ -4,15 +4,15 @@ import { Config } from '../../../build/config'
 import { PageWrapper } from '../../../components'
 import { PagesData } from '../../Page'
 import { getProps } from './props'
-import { BridgesTvlPage } from './view/BridgesTvlPage'
+import { BridgesSummaryPage } from './view/BridgesSummaryPage'
 
-export function getBridgesTvlPage(config: Config, pagesData: PagesData) {
+export function getBridgesSummaryPage(config: Config, pagesData: PagesData) {
   const { props, wrapper } = getProps(config, pagesData)
   return {
     slug: '/bridges/summary',
     page: (
       <PageWrapper {...wrapper}>
-        <BridgesTvlPage {...props} />
+        <BridgesSummaryPage {...props} />
       </PageWrapper>
     ),
   }
