+++ packages/frontend/src/pages/bridges/summary/props/getPageMetadata.ts
