+++ packages/frontend/src/pages/index.ts
@@ -1,7 +1,7 @@
 import { Config } from '../build/config'
 import { getBridgeProjectPages } from './bridges/projects'
 import { getBridgesRiskPage } from './bridges/risk'
-import { getBridgesTvlPage } from './bridges/tvl'
+import { getBridgesSummaryPage } from './bridges/summary'
 import { getDonatePage } from './donate'
 import { getFaqPage } from './faq'
 import { getL2DaysPage } from './l2days'
@@ -10,12 +10,12 @@ import { getMultisigReportDownloadPage } from './multisig-report'
 import { outputPages } from './output'
 import { Page, PagesData } from './Page'
 import { getActivityPage } from './scaling/activity'
-import { getDetailedTvlPage } from './scaling/detailed-tvl'
 import { getLivenessPage } from './scaling/liveness'
 import { getProjectPages } from './scaling/projects'
 import { getProjectTvlBreakdownPages } from './scaling/projects-tvl-breakdown'
 import { getRiskPage } from './scaling/risk'
-import { getTvlPage } from './scaling/tvl'
+import { getTvlPage } from './scaling/summary'
+import { getDetailedTvlPage } from './scaling/tvl'
 
 export async function renderPages(config: Config, pagesData: PagesData) {
   const pages: Page[] = []
@@ -36,7 +36,7 @@ export async function renderPages(config: Config, pagesData: PagesData) {
   pages.push(...getProjectPages(config, pagesData))
   pages.push(...getMetaImagePages(config, tvlApiResponse, activityApiResponse))
 
-  pages.push(getBridgesTvlPage(config, pagesData))
+  pages.push(getBridgesSummaryPage(config, pagesData))
   pages.push(getBridgesRiskPage(config, pagesData))
   pages.push(...getBridgeProjectPages(config, pagesData))
