+++ packages/frontend/src/pages/scaling/summary/index.tsx
@@ -4,15 +4,15 @@ import { Config } from '../../../build/config'
 import { PageWrapper } from '../../../components'
 import { PagesData } from '../../Page'
 import { getProps } from './props'
-import { ScalingDetailedTvlPage } from './view/ScalingDetailedTvlPage'
+import { ScalingSummaryPage } from './view/ScalingSummaryPage'
 
-export function getDetailedTvlPage(config: Config, pagesData: PagesData) {
+export function getTvlPage(config: Config, pagesData: PagesData) {
   const { props, wrapper } = getProps(config, pagesData)
   return {
-    slug: '/scaling/detailedTvl',
+    slug: '/scaling/summary',
     page: (
       <PageWrapper {...wrapper}>
-        <ScalingDetailedTvlPage {...props} />
+        <ScalingSummaryPage {...props} />
       </PageWrapper>
     ),
   }
