+++ packages/frontend/src/pages/scaling/summary/props/index.ts
