+++ packages/frontend/src/pages/scaling/summary/view/FinancialCell.tsx
