+++ packages/frontend/src/pages/scaling/tvl/props/getPageMetadata.ts
@@ -5,7 +5,7 @@ export function getPageMetadata(): PageMetadata {
     title: 'L2BEAT – The state of the layer two ecosystem',
     description:
       'L2BEAT is an analytics and research website about Ethereum layer 2 scaling. Here you will find in depth comparison of major protocols live on Ethereum today.',
-    image: 'https://l2beat.com/meta-images/overview-scaling.png',
-    url: 'https://l2beat.com/scaling/summary',
+    image: 'https://l2beat.com/meta-images/overview-detailed-scaling.png',
+    url: 'https://l2beat.com/',
   }
 }
