+++ packages/frontend/src/pages/scaling/tvl/props/getProps.ts
@@ -1,37 +1,26 @@
 import { Config } from '../../../../build/config'
 import { getFooterProps, getNavbarProps } from '../../../../components'
-import { getChartUrl } from '../../../../scripts/charts/data-controller/ChartDataController'
-import { getTvlWithChange } from '../../../../utils/tvl/getTvlWithChange'
 import { PagesData, Wrapped } from '../../../Page'
-import { TvlPageProps } from '../view/ScalingTvlPage'
+import { ScalingTvlPageProps } from '../view/ScalingTvlPage'
 import { getPageMetadata } from './getPageMetadata'
 import { getScalingTvlView } from './getScalingTvlView'
 
 export function getProps(
   config: Config,
   pagesData: PagesData,
-): Wrapped<TvlPageProps> {
-  const { tvlApiResponse, verificationStatus } = pagesData
-
-  const charts = tvlApiResponse.layers2s
-  const { tvl } = getTvlWithChange(charts)
-
+): Wrapped<ScalingTvlPageProps> {
   return {
     props: {
-      navbar: getNavbarProps(config, 'scaling'),
-      tvlView: getScalingTvlView(
-        config.layer2s,
-        tvlApiResponse,
-        tvl,
-        verificationStatus,
-      ),
-      footer: getFooterProps(config),
       showActivity: config.features.activity,
       showLiveness: config.features.liveness,
-      milestones: config.milestones,
+      navbar: getNavbarProps(config, 'scaling'),
+      footer: getFooterProps(config),
+      detailedTvlView: getScalingTvlView(
+        config.layer2s,
+        pagesData.tvlApiResponse,
+      ),
     },
     wrapper: {
-      preloadApi: getChartUrl({ type: 'layer2-tvl' }),
       metadata: getPageMetadata(),
       banner: config.features.banner,
     },
