+++ packages/frontend/src/pages/scaling/tvl/props/getScalingTvlView.ts
@@ -1,96 +1,73 @@
-import { Layer2, layer2s } from '@l2beat/config'
-import { TvlApiResponse, VerificationStatus } from '@l2beat/shared-pure'
+import { Layer2 } from '@l2beat/config'
+import { TvlApiResponse } from '@l2beat/shared-pure'
 
 import { getIncludedProjects } from '../../../../utils/getIncludedProjects'
 import { orderByTvl } from '../../../../utils/orderByTvl'
-import { getProjectTvlTooltipText } from '../../../../utils/project/getProjectTvlTooltipText'
+import { getTokens } from '../../../../utils/project/getChart'
 import { isAnySectionUnderReview } from '../../../../utils/project/isAnySectionUnderReview'
 import { getRiskValues } from '../../../../utils/risks/values'
-import { getTvlStats, TvlStats } from '../../../../utils/tvl/getTvlStats'
-import { formatPercent, formatUSD } from '../../../../utils/utils'
+import { getDetailedTvlWithChange } from '../../../../utils/tvl/getTvlWithChange'
+import { formatUSD } from '../../../../utils/utils'
 import { ScalingTvlViewEntry } from '../types'
 import { ScalingTvlViewProps } from '../view/ScalingTvlView'
 
 export function getScalingTvlView(
   projects: Layer2[],
   tvlApiResponse: TvlApiResponse,
-  tvl: number,
-  verificationStatus: VerificationStatus,
 ): ScalingTvlViewProps {
-  const included = getIncludedProjects(projects, tvlApiResponse)
-  const ordered = orderByTvl(included, tvlApiResponse)
+  const included = getIncludedProjects(projects, tvlApiResponse).filter(
+    (project) => !project.isLayer3,
+  )
+  const orderedProjects = orderByTvl(included, tvlApiResponse)
+
   return {
-    items: ordered.map((project) =>
-      getScalingTvlViewEntry(
-        project,
-        tvlApiResponse,
-        tvl,
-        verificationStatus.projects[project.id.toString()],
-      ),
+    items: orderedProjects.map((project) =>
+      getScalingTvlViewEntry(tvlApiResponse, project),
     ),
   }
 }
 
 function getScalingTvlViewEntry(
-  project: Layer2,
   tvlApiResponse: TvlApiResponse,
-  aggregateTvl: number,
+  project: Layer2,
   isVerified?: boolean,
 ): ScalingTvlViewEntry {
-  const associatedTokens = project.config.associatedTokens ?? []
-  const apiProject = tvlApiResponse.projects[project.id.toString()]
-
-  let stats: TvlStats | undefined
-
-  if (!apiProject) {
-    if (!project.isUpcoming && !project.isLayer3) {
-      throw new Error(`Project ${project.display.name} is missing in api`)
-    }
-  } else {
-    stats = getTvlStats(apiProject, project.display.name, associatedTokens)
-  }
+  const projectData = tvlApiResponse.projects[project.id.toString()]
+  const charts = projectData?.charts
+  const { parts, partsWeeklyChange } = getDetailedTvlWithChange(charts)
 
   return {
     name: project.display.name,
     slug: project.display.slug,
-    provider: project.display.provider,
     category: project.display.category,
+    provider: project.display.provider,
     riskValues: getRiskValues(project.riskView),
     warning: project.display.warning,
-    isVerified,
     isArchived: project.isArchived,
     showProjectUnderReview: isAnySectionUnderReview(project),
     isUpcoming: project.isUpcoming,
-    isLayer3: project.isLayer3,
-    hostChainName: layer2s.find((l) => l.id === project.hostChain)?.display
-      .name,
-    tvl:
-      stats && escrowsConfigured(project)
-        ? {
-            value: stats.latestTvl,
-            displayValue: formatUSD(stats.latestTvl),
-          }
-        : undefined,
-    tvlTooltip: getProjectTvlTooltipText(project.config),
-    tvlBreakdown:
-      stats && escrowsConfigured(project) ? stats.tvlBreakdown : undefined,
-    oneDayChange:
-      stats && escrowsConfigured(project) ? stats.oneDayChange : undefined,
-    sevenDayChange:
-      stats && escrowsConfigured(project) ? stats.sevenDayChange : undefined,
-    marketShare:
-      stats && escrowsConfigured(project)
-        ? {
-            value: stats.latestTvl / aggregateTvl,
-            displayValue: formatPercent(stats.latestTvl / aggregateTvl),
-          }
-        : undefined,
-    marketShareValue: stats?.latestTvl && stats.latestTvl / aggregateTvl,
-    purpose: project.display.purpose,
+    isVerified,
+    tvl: {
+      value: parts.tvl,
+      displayValue: formatUSD(parts.tvl),
+    },
+    cbv: {
+      value: parts.canonical,
+      displayValue: formatUSD(parts.canonical),
+    },
+    ebv: {
+      value: parts.external,
+      displayValue: formatUSD(parts.external),
+    },
+    nmv: {
+      value: parts.native,
+      displayValue: formatUSD(parts.native),
+    },
+    tvlChange: partsWeeklyChange.tvl,
+    ebvChange: partsWeeklyChange.external,
+    cbvChange: partsWeeklyChange.canonical,
+    nmvChange: partsWeeklyChange.native,
+    tokens: getTokens(project.id, tvlApiResponse, true),
     stage: project.stage,
   }
-}
-
-export function escrowsConfigured(project: Layer2) {
-  return project.config.escrows.length > 0
 }
