+++ packages/frontend/src/pages/scaling/tvl/types.ts
@@ -1,6 +1,6 @@
 import { Layer2, StageConfig } from '@l2beat/config'
 
-import { TVLBreakdownProps } from '../../../components/TVLBreakdown'
+import { TokenControl } from '../../../components/chart/TokenControls'
 import { RiskValues } from '../../../utils/risks/types'
 import { ValueWithDisplayValue } from '../../types'
 
@@ -16,14 +16,13 @@ export interface ScalingTvlViewEntry {
   showProjectUnderReview?: boolean
   isUpcoming?: boolean
   tvl?: ValueWithDisplayValue
-  isLayer3?: boolean
-  hostChainName?: string
-  tvlTooltip?: string
-  tvlBreakdown?: TVLBreakdownProps
-  oneDayChange?: string
-  sevenDayChange?: string
-  marketShare?: ValueWithDisplayValue
-  marketShareValue?: number
-  purpose: string
+  cbv?: ValueWithDisplayValue
+  ebv?: ValueWithDisplayValue
+  nmv?: ValueWithDisplayValue
+  tvlChange?: string
+  ebvChange?: string
+  cbvChange?: string
+  nmvChange?: string
+  tokens: TokenControl[]
   stage: StageConfig
 }
